syscall: use size_t/ssize_t for read/write count and return types
Change sys_read, sys_write, kern_read, kern_write from int to size_t
for the count/length parameter and ssize_t for the return type. This
aligns the kernel interface with POSIX semantics where counts are
unsigned sizes and returns are signed to accommodate error codes.

Updated across all layers:
- Kernel implementation (sys/kern/syscall.c)
- Extern declarations (sys/include/sys/syscall_impl.h)
- Internal declarations (sys/include/sys/kern_syscalls.h)
- ELKS personality wrappers (sys/exec/perso/perso_elks.c)

LibC wrappers in lib/c/src/sys.c already used size_t/ssize_t.

+++ sys/include/sys/kern_syscalls.h
@@ -25,8 +25,8 @@ struct sigaction;
 /* Internal kernel versions of syscalls that take kernel pointers */
 int kern_acct(const char *path);
 int kern_open(const char *path, int flags, int mode);
-int kern_read(int fd, char *buf, int len);
-int kern_write(int fd, const char *buf, int len);
+ssize_t kern_read(int fd, char *buf, size_t len);
+ssize_t kern_write(int fd, const char *buf, size_t len);
 int kern_close(int fd);
 int kern_stat(const char *path, struct stat *buf);
 int kern_lstat(const char *path, struct stat *buf);
